convert daxpy example to alpaka
