+++ CMakeLists.txt
@@ -86,7 +86,6 @@ if (LLAMA_BUILD_EXAMPLES)
 	add_subdirectory("examples/bytesplit")
 	add_subdirectory("examples/bitpackint")
 	add_subdirectory("examples/bitpackfloat")
-	add_subdirectory("examples/daxpy")
 
 	# alpaka examples
 	find_package(alpaka 0.7.0 QUIET)
@@ -95,6 +94,7 @@ if (LLAMA_BUILD_EXAMPLES)
 		add_subdirectory("examples/alpaka/vectoradd")
 		add_subdirectory("examples/alpaka/asyncblur")
 		add_subdirectory("examples/alpaka/pic")
+		add_subdirectory("examples/alpaka/daxpy")
 	elseif()
 		message(WARNING "Could not find alpaka. Alpaka examples are disabled.")
 	endif()
