+++ examples/alpaka/daxpy/CMakeLists.txt
@@ -1,15 +1,15 @@
 cmake_minimum_required (VERSION 3.15)
-project(llama-daxpy CXX)
+project(llama-alpaka-daxpy CXX)
 
 #find_package(Vc QUIET)
 find_package(OpenMP REQUIRED)
 if (NOT TARGET llama::llama)
 	find_package(llama REQUIRED)
 endif()
-
-add_executable(${PROJECT_NAME} daxpy.cpp)
+find_package(alpaka 0.7.0 REQUIRED)
+alpaka_add_executable(${PROJECT_NAME} daxpy.cpp ../../common/Stopwatch.hpp ../../common/hostname.hpp)
 target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_17)
-target_link_libraries(${PROJECT_NAME} PRIVATE llama::llama OpenMP::OpenMP_CXX)
+target_link_libraries(${PROJECT_NAME} PRIVATE llama::llama OpenMP::OpenMP_CXX alpaka::alpaka)
 
 if (MSVC)
 	target_compile_options(${PROJECT_NAME} PRIVATE /arch:AVX2 /fp:fast)
